+++ yarn-project/circuit-types/src/logs/function_l2_logs.ts
@@ -7,7 +7,6 @@ import { randomBytes } from 'crypto';
 import { LogType } from './log_type.js';
 import { UnencryptedL2Log } from './unencrypted_l2_log.js';
 
-// ???????????????????????????????????
 /**
  * Data container of logs emitted in 1 function invocation (corresponds to 1 kernel iteration).
  */
