+++ yarn-project/end-to-end/scripts/docker-compose.yml
@@ -25,6 +25,7 @@ services:
       WS_BLOCK_CHECK_INTERVAL_MS: 50
       PXE_BLOCK_POLLING_INTERVAL_MS: 50
       ARCHIVER_VIEM_POLLING_INTERVAL_MS: 500
+      AVM_ENABLED: ${AVM_ENABLED:-}
     ports:
       - '8080:8080'
